chore: fix annoying warning

+++ da/da.c
@@ -23,7 +23,7 @@ int64_t da_indexof_impl(void* arr, void* elem, size_t elem_size) {
     if (!arr) return -1;
     da_header_t* header = (da_header_t*)arr - 1;
     char* byte_arr = (char*)arr;
-    for (int64_t i = 0; i < header->size; ++i) {
+    for (int64_t i = 0; i < (int64_t)(header->size); ++i) {
         if (memcmp(&byte_arr[i*elem_size], elem, elem_size) == 0) return i;
     }
     return -1;
